+++ lib/calc.h
@@ -1,9 +1,9 @@
 #ifndef __CALC__
 #define __CALC__
 
-	int sum(int num1, int num2);
-	int multi(int num1, int num2);
-	int subtraction(int num1, int num2);
-	int division(int num1, int num2);
+	float sum(float num1, float num2);
+	float multi(float num1, float num2);
+	float subtraction(float num1, float num2);
+	float division(float num1, float num2);
 
 #endif
